chore: move react-docs file
